service/ai_assistant/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiomysql==0.3.0
redis==5.2.1
python-jose[cryptography]==3.4.0
cryptography<45.0.0
passlib[bcrypt]==1.7.4
pycryptodome==3.21.0
pydantic-settings==2.6.1
python-multipart==0.0.22
httpx==0.28.1
dashscope==1.25.12
alibabacloud-bailian20231229==2.8.1
alibabacloud-tea-openapi>=0.3.14
alibabacloud-tea-util>=0.3.13
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
Pillow>=10.0.0
langchain-core==0.3.51
loguru==0.7.2